Further changes to traits and abilities
Magic Missile now has scaling instead of the old system, 200% of base damage. and now counts as both projectile and explosive damage.

New traits added that give the mobs stack of corresponding blood effects (brutal, relentless, etc) when they cast an ability. The trait gives the mob 2 stacks. Healer and Augmenter apply these to their targets like with some other traits.

Standardized the look of projectiles and raycasts with traits in Mob Abilities. Each trait will now give off the consistent particles in the animation so it is more clear what type of attack it is via the raycast as well as the mob.