Retooled Webshot to Quickdraw
Webshot no longer exists, replaced with Quickdraw

Quickdraw is another preparation skill like Sweep which attacks very quickly.

A raycast is drawn that the player has a half second to move out of or they take 200% damage.

Mob is knocked back from attack slightly.

Also features optional tag "warpshot" that allows the mob to teleport instead of being knocked back.

New animation for Sweep and Quickdraw that plays when the mobs are tokened. (instead of the constant anger particles)

Needle now has the preparation animation and is classified similar to Sweep in that regard.