Added ca.soul_chain tags
ca.soul_chain_start is used on the bottom mob of a stack, ca.soul_chain is used on the rest of the stack

Mobs when spawned in are stack analyzed and all given an id. When one mob with an id dies, all other mobs with said id die.