Beginning of the Tag Implementation
Adding in event tags

These are function tags that can be extended off of and allow mapmakers to hook their own datapacks into Cartographer effects.

Currently the 3 that exist in this push are Adrenaline triggers, Frenzy triggers, and Energetic triggers. Currently, the functions in these are the debug messages so the tags do not harm performance.