Stealthed as a Custom Statuses
Stealthed is a custom status, and functions identically to the player in Concealed from the same enchantment.

Stealth is a potion effect of Invisibility 2 or higher and makes you invisible to mobs that are more than 2 blocks away from you.

Stealthed is lost when you take mob damage or deal damage to a mob, or when the duration runs out.

Stealthed does NOT grant a 20% damage bonus like Concealed.