Overcharge Cauterize and Singe changes
- Fixed a bug with Trueshot and Explosive that caused the resulting explosion to do less damage than it should have.

- Fixed a bug with Overcharge and Explosive that caused the resulting explosion to do less damage than it should have.

Overcharge has been nerfed and rescaled:

- Overcharge now scales 6 tiers instead of 3. The tiers gain 50% faster, (0.5 seconds per tier instead of 0.75)

- Each tier overall is a 17% damage boost, rather than a 50% one.
(This means that max tiers of overcharge will double the damage of the arrow.)

(the tldr is overcharge takes additional time to ramp up to max and deals less increased damage. However it will reach lower levels of power faster. These nerfs are to combat the synergy of how powerful Overcharge is with Power, to avoid having to remove the Power synergy outright. This makes an Overcharge 3 and Power 3 bow roughly equivalent in damage. It is also being done to make the enchantment feel more responsive while drawing back.)

Cauterize has been simplified.
- Scales up to level 10.

- Reworked effect: Cauterize instantly extinguishes a mob on fire, dealing 15% (per level) of the remaining fire tick as burst damage. Cauterize will always deal at least 1 damage.

- Maxes at 10 levels so max level is 150% of the fire tick remaining.

- Cauterize will deal Sharpness equivalent damage to mobs that are fire immune instead (based on level).

Singe nerfs:
- Singe will deal Sharpness equivalent damage instead of Smite equivalent damage.
- Singe now caps at 10 (was 5.)